requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-bootstrap-v5==1.0.11
django-bootstrap4==22.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-emoji-picker==0.0.6
django-extensions==3.1.5
django-heroku==0.3.1
django-phone-field==1.8.1
django-tinymce==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
oauthlib==3.2.0
paramiko==2.11.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
storage==0.0.4.3
urllib3==1.26.9
whitenoise==6.2.0
